+++ src/components/dialog/dialog.pw.tsx
@@ -10,6 +10,7 @@ import {
   TopModalOverride,
   DialogWithAutoFocusSelect,
   DialogWithStepSequence,
+  DialogComponentFocusableSelectors,
 } from "./components.test-pw";
 
 import {
@@ -29,6 +30,7 @@ import {
   checkAccessibility,
   getStyle,
   waitForAnimationEnd,
+  waitForElementFocus,
 } from "../../../playwright/support/helper";
 import { CHARACTERS, SIZE } from "../../../playwright/support/constants";
 import { getDataElementByValue } from "../../../playwright/components";
@@ -381,6 +383,30 @@ test.describe("Testing Dialog component properties", () => {
     await dialog.press("Shift+Tab");
     await dialog.press("Shift+Tab");
     await expect(select).toBeFocused();
+  });
+
+  test("should render component with first input and button as focusableSelectors", async ({
+    mount,
+    page,
+  }) => {
+    await mount(<DialogComponentFocusableSelectors />);
+
+    const dialogElement = page.getByRole("dialog");
+    const focusedElement = page.locator("*:focus");
+    const firstInputElement = page.getByLabel("First name");
+    const secondInputElement = page.getByLabel("Surname");
+    const openToastElement = getDataElementByValue(page, "open-toast");
+
+    await waitForElementFocus(page, dialogElement);
+    await dialogElement.press("Tab");
+    await focusedElement.press("Tab");
+
+    await expect(firstInputElement).toBeFocused();
+
+    await focusedElement.press("Tab");
+
+    await expect(secondInputElement).not.toBeFocused();
+    await expect(openToastElement).toBeFocused();
   });
 });